requirements.txt
pyusb
lz4
